fix: standardize all timestamps to i64
Unify timestamp types across the codebase to i64, eliminating mixed i64/u64 usage that caused unsafe casts and defensive filtering.  - Convert Nostr SDK u64 timestamps to i64 at API boundary - Update AdminChatLastSeen, AdminChatUpdate, OrderMessage,   MessageNotification structs to use i64 - Simplify helpers by removing negative timestamp filtering - Use entry API to ensure last_seen updates persist for new disputes - Remove unnecessary casts now that types are consistent

+++ src/ui/tab_content.rs
@@ -163,7 +163,7 @@ pub fn render_messages_tab(
                 "Order: Unknown".to_string()
             };
 
-            let timestamp = DateTime::<Utc>::from_timestamp(msg.timestamp as i64, 0)
+            let timestamp = DateTime::<Utc>::from_timestamp(msg.timestamp, 0)
                 .map(|dt| dt.format("%Y-%m-%d %H:%M:%S").to_string())
                 .unwrap_or_else(|| "Unknown time".to_string());
